hugeadm: Use unsigned long long when adjusting the pool page sizes
When specifying the amount of memory to allocate as huge pages with
--pool-pages-max, it is possible to overflow the huge page count on 32 bit
systems with large amounts of memory. For example, specifying --pool-pages-max
DEFAULT:8G will overflow long and set the wrong value.

This patch converts size_to_smaller_units() to unsigned long long to manage
the overflow.  The existing callers of size_to_smaller_units() should be
all right as they are always talking about the context of a huge page size
which is never going to overflow the long type.

Tested on 32-bit X86.

+++ hugeadm.c
@@ -1004,7 +1004,7 @@ enum {
 
 static long value_adjust(char *adjust_str, long base, long page_size)
 {
-	long adjust;
+	unsigned long long adjust;
 	char *iter;
 
 	/* Convert and validate the adjust. */
